requirements.txt
flask
google-generativeai
scikit-learn
pandas
matplotlib
streamlit